feat: extract http api from stores
抽离独立的http api, 不应该和stores混在一起

Log:
