+++ package.json
@@ -53,6 +53,7 @@
 		"@tarojs/runtime": "3.6.20",
 		"@tarojs/shared": "3.6.20",
 		"@tarojs/taro": "3.6.20",
+		"@vueuse/core": "^10.7.0",
 		"axios": "^1.5.0",
 		"dayjs": "^1.11.10",
 		"pinia": "^2.1.6",
