+++ src/api/index.ts
@@ -1,0 +1,9 @@
+export * from "./account";
+export * from "./agreement";
+export * from "./config";
+export * from "./http";
+export * from "./post-create";
+export * from "./post-list";
+export * from "./thread-index";
+export * from "./thread-info";
+export * from "./thread-search";
