+++ src/api/thread-info.ts
@@ -1,114 +1,9 @@
-import { defineStore } from "pinia";
-import { ref } from "vue";
+import { http } from "./http";
 
-import { apiServer, http } from "./http";
-
-export const useThreadStore = defineStore("thread", () => {
-  const page = ref(1);
-  const loaded = ref(false);
-  const item = ref(null as ThreadInfoResponse["data"] | null);
-  const posts = ref([] as Post[]);
-  const postLoaded = ref(false);
-  const postCount = ref(0);
-  const postLimit = ref(10);
-  const server = ref(apiServer);
-
-  async function thraadInfo(id: number) {
-    loaded.value = false;
-    const info = await getThraadInfo(id);
-    item.value = info.data.data;
-    loaded.value = true;
-  }
-  async function threadPost(id: number, offset: number, limit: number) {
-    postLoaded.value = false;
-    const threads = await getThreadPost(id, { offset, limit });
-    posts.value = threads.data.data;
-    postCount.value = threads.data.total_count;
-    postLimit.value = threads.data.total_limit;
-    postLoaded.value = true;
-  }
-  async function load(id: number) {
-    page.value = 1;
-    await thraadInfo(id);
-    threadPost(id, 0, postLimit.value);
-  }
-  function pageChange(p: number) {
-    if (item.value) {
-      page.value = p;
-      postLoaded.value = false;
-      threadPost(item.value.id, p - 1, postLimit.value);
-    }
-  }
-  function createPost(req: CreatePostRequest) {
-    return createThreadPost(req);
-  }
-  return {
-    page,
-    loaded,
-    item,
-    postCount,
-    postLimit,
-    posts,
-    postLoaded,
-    server,
-    load,
-    pageChange,
-    createPost,
-  };
-});
-
-async function getThraadInfo(id: number) {
+export async function ThraadInfo(id: number) {
   return http.get<ThreadInfoResponse>("/api/v1/thread/info", {
     params: { id },
   });
-}
-async function getThreadPost(
-  id: number,
-  opt: {
-    offset?: number;
-    limit?: number;
-  }
-) {
-  const params = {
-    id,
-    offset: 0,
-    limit: 10,
-    ...opt,
-  };
-  return http.get<PostListResponse>("/api/v1/thread/post/list", {
-    params,
-  });
-}
-
-async function createThreadPost(body: CreatePostRequest) {
-  return http
-    .post<CreatePostResponse>("/api/v1/post", body, {
-      params: { captcha_id: body.captcha_id },
-    })
-    .then((resp) => {
-      if (resp.data.code != 0) {
-        throw resp.data.msg;
-      }
-      return resp;
-    });
-}
-
-interface CreatePostRequest {
-  captcha_id: string;
-  validate: string;
-  forum_id: number;
-  thread_id: number;
-  message: string;
-  quote_user_id: number;
-  user_ids?: number[];
-}
-interface CreatePostResponse {
-  code: number;
-  data: {
-    id: number;
-    offset: number;
-  };
-  msg: string;
 }
 
 interface ThreadInfoResponse {
