+++ src/assets/captcha.js
